+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - `./src/modules/wordle.js` new game function to work with first new command
 - New game function in wordle to work better with interaction
 
+### Removed
+- Help embed
+
 ## [Alpha 1.2.1](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
 ### Added
 - Help embed