+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog (Alpha Versions)
-## [Release Alpha 1.3](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
+## [Release Alpha 1.3](https://github.com/Nevah5/wordlebot/releases/tag/a1.3) - 06.02.2022
 ### Added
 - Guild commands for testing
 - Guess command